chore(coordinator): refactor aggregation target block number handling for clarity and consistency
